Enforce tables

+++ example/09_live_shader/09_live_shader.cpp
@@ -87,7 +87,7 @@ void ex_09_live_shader(Shell& app, Widget& parent, Dockbar& dockbar)
 		TextEdit& text_edit = ui::code_edit(*edit.m_body, source, 0);
 		if(text_edit.m_changed)
 		{
-			program.m_sources[size_t(ShaderType::Fragment)] = source.c_str();
+			program.m_sources[ShaderType::Fragment] = source.c_str();
 			program.reload();
 		}
 		text_edit.m_changed = false;
