+++ src/tool/Tools/ScaleTool.cpp
@@ -62,7 +62,7 @@ namespace mud
 	Item& scale_2d_gizmo(Gnode& parent, Axis axis, Colour colour, uint32_t flags = 0U)
 	{
 		Gnode& node = gfx::transform(parent, {}, 0.5f * to_vec3(axis), ZeroQuat);
-		return gfx::shape(node, Quad(0.2f, c_tangents[uint(axis)], c_binormals[uint(axis)]), Symbol(colour, Colour::None, true, true), flags);
+		return gfx::shape(node, Quad(0.2f, c_tangents[axis], c_binormals[axis]), Symbol(colour, Colour::None, true, true), flags);
 	}
 
 	Item& scale_3d_gizmo(Gnode& parent, Colour colour, uint32_t flags = 0U)
